+++ example/android/app/src/main/kotlin/dev/simonit/fluda/example/MainActivity.kt
@@ -1,6 +1,0 @@
-package dev.simonit.fluda.example
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity() {
-}